Move square_brackets function to its own file
